// Frontend/src/app/core/models/customer.ts
export interface Customer {
  customerId: number;
  customerName: string | any;
  mobileNo: string | any;
  email: string | any;
  address: string | any;
  city: string | any;
  state: string | any;
  gstIn: string | any;
  createdDate: Date | any;
  updatedDate: Date | any;
  isActive: boolean | any;
}